Change maxTime parameter type from int to long.

+++ src/main/com/mongodb/DBCursor.java
@@ -185,7 +185,7 @@ public class DBCursor implements Iterator<DBObject> , Iterable<DBObject>, Closea
      * @return same DBCursor for chaining operations
      * @since 2.12.0
      */
-    public DBCursor maxTime(final int maxTime, final TimeUnit timeUnit) {
+    public DBCursor maxTime(final long maxTime, final TimeUnit timeUnit) {
         _maxTimeMS = MILLISECONDS.convert(maxTime, timeUnit);
         return this;
     }
